Extract payload operations into payload module
